% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshr.R
\name{freshr}
\alias{freshr}
\title{Unload none-base packages and remove all objects from environment}
\usage{
freshr()
}
\value{
This function returns no value. It is called for its side effect
of cleaning up the global environment at script's start up/
}
\description{
Ever experienced a time when you pass an R script to your
colleague, and it does not work on their machine because you forgot to
add a library import that you did interactively? freshr is here to
help! By calling freshr::freshr() at the top of your every script, you
create a fresh environment for you to test reproducibility before you
send your script to someone else.
}
